+++ src/main/java/woowacourse/shoppingcart/dto/cart/request/CartRemovalRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.cart.request;
 
 import java.util.List;
