+++ src/main/java/woowacourse/shoppingcart/dto/cart/request/CartRequest.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.cart.request;
 
 public class CartRequest {
