+++ src/main/java/woowacourse/shoppingcart/dto/cart/response/CartsResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.cart.response;
 
 import java.util.List;
