+++ src/main/java/woowacourse/shoppingcart/dto/customer/response/CustomerResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.customer.response;
 
 public class CustomerResponse {
