+++ src/main/java/woowacourse/shoppingcart/dto/product/response/ProductsResponse.java
@@ -1,4 +1,4 @@
-package woowacourse.shoppingcart.dto;
+package woowacourse.shoppingcart.dto.product.response;
 
 import java.util.List;
